Change container stopping and starting actions to 'replace'
This changes the stop mode of the container. When in 'fail', the stop
job would fail if a start job was already in queue. This occured when
the client had just been started and an update on this container was
already planned on the server - the client then wants to stop container
before updating it but the container is still starting.
This 'replace' stop mode prevents this from happening, as it replaces
the start job by the stop job.

For the same reason, the update sometimes fails because a
stop job is running, thus making the start job impossible.