Add a thread to feedback the client the status of notify containers
The purpose of this thread is, for notify containers, to feedback the
result of their execution. It will wait on a Unix socket for a message
coming from the container itself. Depending of the datagram information,
it feedbacks the server either positively, or negatively - along with
some information about the service status.

If the container has failed, the thread will attempt to rollback the
container to its previous version (if there is one).

If, for some reason, the client never receives the datagram, the socket
will timeout and consider the deployment as failed and once again
attempt to rollback. This timeout is defined in the container's recipe
and read from hawkbit's metadata.

The thread is started just before the service, after having downloaded
the necessary objects to update. This way there is a minimal time before
the thread and the container start, which important since the thread can
timeout.

Also, two methods were added (set_current_revision and get_previous_rev)
to read the parent revision and write the current revision of the notify
container. This is because there is no way with the OSTree python API to
get the parent commit id of the current revision, needed in case of a
rollback.