Cleanup the sysroot on client initialization
Instead of when the client has just staged the new tree.

Since the new tree is only staged but not deployed, the cleanup has no
effect. Instead, the cleanup should be done once the tree is deployed -
in other words, when the system has rebooted.

Doing a cleanup on first initialization of the target, or on reboot
(with no updates) does not have any side-effect.