Correct the feedback process for notify containers
We now use 1 Unix Socket per feedback thread, knowing that each feedback thread is associated with a unique notify container.
It allows a better monitoring of transactions bewteen notify containers (senders) and feedback threads (receivers).

Moreover, we now start all containers at the same time : so everything (= start + acknoledgement of the success or failure of container startup) fits in the 35s during which the connexion is opened between the feedback thread and the socket.